[#53740] Upgrade/limit C# language version to 7.2
That's an upgrade for .NET Framework projects which had a C# 7.0
limit. .NET6 projects had no limit and are limited for uniformity.

While Renode can be built for .NET6 with C# 7.3+ features, the same
code needs to be buildable for .NET Framework which only supports
C# versions up to 7.3.

Also, it must be buildable with the Debian/Ubuntu Mono packages
which don't have `csc`, i.e. Roslyn compiler, packaged. The old C#
compiler, `mcs`, supports C# versions only up to 7.2.